SEQNG-272 Let the web client module pass scalastyle checks #time 2h
