+++ modules/edu.gemini.seqexec.web/edu.gemini.seqexec.web.client/src/main/scala/edu/gemini/seqexec/web/client/semanticui/align.scala
@@ -1,0 +1,16 @@
+package edu.gemini.seqexec.web.client.semanticui
+
+import scalaz.Equal
+
+sealed trait Aligned
+
+object Aligned {
+  case object None extends Aligned
+  case object Left extends Aligned
+  case object Center extends Aligned
+  case object Right extends Aligned
+
+  implicit val equal: Equal[Aligned] = Equal.equalA[Aligned]
+}
+
+trait SemanticUIAlign extends Aligned
