+++ modules/edu.gemini.seqexec.web/edu.gemini.seqexec.web.client/src/main/scala/edu/gemini/seqexec/web/client/semanticui/elements/table/TableHeader.scala
@@ -16,7 +16,7 @@ object TableHeader {
   )
 
   object Props {
-    def zero = Props()
+    def zero: Props = Props()
   }
 
   private val component = ScalaComponent.builder[Props]("th")
