+++ modules/edu.gemini.seqexec.web/edu.gemini.seqexec.web.client/src/main/scala/edu/gemini/seqexec/web/client/semanticui/package.scala
@@ -2,7 +2,7 @@ package edu.gemini.seqexec.web.client
 
 import japgolly.scalajs.react.vdom.html_<^.VdomAttr
 
-package object semanticui {
+package object semanticui extends SemanticUISize with SemanticUIWidth with SemanticUIAlign {
   // Custom attributes used by SemanticUI
   val dataTab     = VdomAttr("data-tab")
   val dataTooltip = VdomAttr("data-tooltip")
